Time: 3 ms (13.67%), Space: 8.5 MB (84.49%) - LeetHub

+++ 3870-count-commas-in-range/3870-count-commas-in-range.cpp
@@ -1,14 +1,13 @@
 class Solution {
 public:
-    long long countCommas(int n) {
-        long long count = 0;
+    int countCommas(int n) {
+        long long count=0;
 
-        for (int i = 999; i <= n; i++) {
-            if (i >= 1000) {
+        for(int i=999;i<=n;i++){
+            if(i>=1000){
                 count++;
             }
         }
-
         return count;
     }
 };
